requirements.txt
numpy==1.24.2
pandas==1.5.3
protobuf==3.20.3
streamlit==1.31.1
google-cloud-bigquery==3.18.0
joblib==1.3.2